Enabled -Wall -Wextra -pedantic, then fixed all errors that occurred.
- Many non-void functions ended without a return statement.
- Many comparisons between signed and unsigned integers.
- Many TDetector subclasses did not define all of the abstract methods in TDetector.
- TS800Scaler.h was missing the ifdef guards
- TUnpackingLoop::HandleGEBMode3 did not use the split up fragments,
    instead passing the original event in multiple times.

+++ include/TWriteLoop.h
@@ -66,7 +66,7 @@ private:
   std::vector<TUnpackedEvent*> learning_queue;
   std::atomic_int learning_phase_size;
   std::atomic_bool in_learning_phase;
-  int learning_phase_length;
+  size_t learning_phase_length;
 #endif
 
   ClassDef(TWriteLoop, 0);
